function-app/requirements.txt
azure-functions
azure-storage-blob
pandas
numpy
openpyxl
pyodbc
sqlalchemy